+++ test/cases/specific_schema_test_sqlserver.rb
@@ -107,6 +107,21 @@ class SpecificSchemaTestSQLServer < ActiveRecord::TestCase
     assert_equal r, SSTestEdgeSchema.where('crazy]]quote' => 'crazyqoute').first
   end
 
+  it 'various methods to bypass national quoted columns for any column, but primarily useful for char/varchar' do
+    value = Class.new do
+      def quoted_id
+        "'T'"
+      end
+    end
+    # Using ActiveRecord's quoted_id feature for objects.
+    assert_sql(/@0 = 'T'/) { SSTestDatatypeMigration.where(char_col: value.new).first }
+    assert_sql(/@0 = 'T'/) { SSTestDatatypeMigration.where(varchar_col: value.new).first }
+    # Using our custom char type data.
+    data = ActiveRecord::Type::SQLServer::Char::Data
+    assert_sql(/@0 = 'T'/) { SSTestDatatypeMigration.where(char_col: data.new('T')).first }
+    assert_sql(/@0 = 'T'/) { SSTestDatatypeMigration.where(varchar_col: data.new('T')).first }
+  end
+
   # With column names that have spaces
 
   it 'create record using a custom attribute reader and be able to load it back in' do